requirements.txt
ipywidgets
notebook
ensure
pyyaml
types-PyYAML
# py-youtube